+++ libft/src/printf.c
@@ -11,6 +11,7 @@
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
+#include <stdarg.h>
 
 int		format_string(va_list arg, t_format fx, int fd)
 {
@@ -49,7 +50,7 @@ int		format_string_u(va_list arg, t_format fx, int fd)
 	empty = (wchar_t)(fx.flag[3] == '0' && fx.flag[0] != '-' ? L'0' : L' ');
 	if (!(tmp = va_arg(arg, int *)))
 		tmp = L"(null)";
-	if ((MB_CUR_MAX < ft_charlen(*tmp)))
+	if ((MB_CUR_MAX < (size_t)ft_charlen(*tmp)))
 		return (-1);
 	if (fx.precs >= 0)
 		tmp = ft_strndup_u(tmp, fx.precs);
